# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/availabilities": {
            "get": {
                "description": "Get availabilities by single date or date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availability"
                ],
                "summary": "Get availabilities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Capability",
                        "name": "Capability",
                        "in": "header"
                    },
                    {
                        "description": "Request Payload",
                        "name": "AvailabilityPayload_Rq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AvailabilityPayload_Rq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.AvailabilityPayload_Rs_Pricing"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/availabilities/add": {
            "post": {
                "description": "Adds availabilities for a product within a single date or date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "availability"
                ],
                "summary": "Add availabilities",
                "parameters": [
                    {
                        "description": "Request Payload for Adding Availabilities",
                        "name": "AvailabilityNewPayload_Rq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AvailabilityNewPayload_Rq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AvailabilityNewPayload_Rq": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "localDate": {
                    "type": "string"
                },
                "localDateEnd": {
                    "type": "string"
                },
                "localDateStart": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "string"
                }
            }
        },
        "model.AvailabilityPayload_Rq": {
            "type": "object",
            "properties": {
                "localDate": {
                    "type": "string"
                },
                "localDateEnd": {
                    "type": "string"
                },
                "localDateStart": {
                    "type": "string"
                }
            }
        },
        "model.AvailabilityPayload_Rs_Pricing": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "localDate": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "productName": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "vacancies": {
                    "type": "integer"
                }
            }
        },
...
    }
}